updates for PR #176
