+++ src/Automation/CSE.Automation/Interfaces/ICosmosDBSettings.cs
@@ -1,0 +1,9 @@
+﻿namespace CSE.Automation.Interfaces
+{
+    interface ICosmosDBSettings : ISettingsValidator
+    {
+        string Uri { get; }
+        string Key { get; }
+        string DatabaseName { get; }
+    }
+}
