+++ src/Automation/CSE.Automation/Interfaces/ISettingsValidator.cs
@@ -1,0 +1,8 @@
+﻿namespace CSE.Automation.Interfaces
+{
+    public interface ISettingsValidator
+    {
+        void Validate();
+    }
+
+}
